se crea medoto actualizar datagrodview al editar o guardar un producto o persona

+++ Biblioteca/Producto.cs
@@ -12,7 +12,7 @@ namespace Biblioteca
         public string codigoProducto { get; set; }
         public string nombreProducto { get; set; }
         public string descripcionProducto { get; set; }
-        public int precioProducto { get; set; }
+        public long precioProducto { get; set; }
         public int stockActualProducto { get; set; }
         public int stockCriticoProducto { get; set; }
         public int idTipoProducto { get; set; }
